Time: 235 ms (19.73%), Space: 37 MB (23.79%) - LeetHub

+++ 1383-maximum-performance-of-a-team/1383-maximum-performance-of-a-team.cpp
@@ -20,6 +20,6 @@ public:
             }
             ans=max(ans,sum*x);
         }
-        return ans%(int) mod;
+        return ans% mod;
     }
 };
